refactor(kaylendog/nodes): split packet into sepparate file
