+++ crates/comm/src/socket/mod.rs
@@ -1,18 +1,17 @@
 //! Defines the UDP socket abstraction and first-layer packet format used for communication between peers.
 
+pub mod error;
+mod packet;
+
 use std::{
-    cmp::Ordering,
     collections::{hash_map::Entry, HashMap},
-    io::{self, Cursor, Read, Write},
+    io::{Read, Write},
     net::SocketAddr,
     sync::Arc,
 };
 
-use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
-use flate2::{read::GzDecoder, write::GzEncoder, Compression};
-use protocol::{prost::DecodeError, try_decode_packet, MessageType, ProtocolPacket};
+use protocol::{MessageType, ProtocolPacket};
 use rand::{rngs::OsRng, seq::IteratorRandom};
-use thiserror::Error;
 use tokio::{
     net::UdpSocket,
     sync::{mpsc, RwLock},
@@ -22,19 +21,14 @@ use tracing::{debug, error, span, trace};
 use crate::{
     crypto::{Crypto, DoubleRatchet},
     maybe_continue,
-    peer::{Peer, PeerError, PeerState},
+    peer::{Peer, PeerState},
     try_break, try_continue,
 };
 
-/// The magic number used to identify packets sent over the network.
-pub const SOCKET_PACKET_MAGIC_NUMBER: u32 = 0x010203;
-
-/// The minimum size of an encoded [SocketPacket].
-// 3 (Magic) + 1 (Packet type) + 4 (Packet number) + 4 (Chunk number) + 4 (Compressed Data length) + 4 (Data length)
-pub const MIN_SOCKET_PACKET_SIZE: usize = 3 + 1 + 4 + 4 + 4 + 4;
-
-/// The maximum size of a UDP datagram.
-pub const UDP_MAX_DATAGRAM_SIZE: usize = 65_507;
+use self::{
+    error::SocketError,
+    packet::{SocketPacket, UDP_MAX_DATAGRAM_SIZE},
+};
 
 /// Number of peers to send gossip to
 const GOSSIP_COUNT: usize = 3;
@@ -50,61 +44,6 @@ pub struct Socket {
     pub crypto: Arc<RwLock<Crypto>>,
     /// Username used to identify this current node
     pub username: String,
-}
-
-/// An enumeration of possible errors that can occur when working with [Socket].
-#[derive(Error, Debug)]
-pub enum SocketError {
-    /// An unknown error occurred.
-    #[error("Unknown error")]
-    Unknown,
-    /// A connection to a peer timed ou.t
-    #[error("Connection timed out")]
-    ConnectionTimeout,
-    /// A connection to a peer already exists.
-    #[error("Already connected")]
-    ConnectionExists,
-    /// A connection to a peer is dead.
-    #[error("Not connected")]
-    ConnectionDead,
-    /// An IO operation failed.
-    #[error("Encountered an IO error")]
-    IoError(#[from] std::io::Error),
-    /// A packet failed to encode.
-    #[error("Failed to encode packet")]
-    EncodeFail(#[from] protocol::prost::EncodeError),
-    /// A peer operation failed.
-    #[error("Failed to process peer operation")]
-    PeerError(#[from] PeerError),
-    /// Trying to start a ratchet when it exists
-    #[error("Ratchet exists")]
-    RatchetExists,
-    /// Tried to send gossip, but 0 peers connected
-    #[error("No peer for gossip")]
-    NoPeer,
-}
-
-/// An enumeration of possible errors that can occur when working with [ProtocolPacket]s.
-#[derive(Error, Debug)]
-pub enum SocketPacketDecodeError {
-    /// An unknown error occurred.
-    #[error("Unknown error")]
-    Unknown,
-    /// The magic number in the packet was incorrect.
-    #[error("Magic number incorrect")]
-    BadMagic,
-    /// An unknown packet type was encountered.
-    #[error("Unknown packet type")]
-    BadPacketType,
-    /// The packet was too small to be valid.
-    #[error("Packet too small")]
-    BadSize,
-    /// An IO operation failed.
-    #[error("Encountered an IO error")]
-    IoError(#[from] std::io::Error),
-    /// Packet decoding failed.
-    #[error("Failed to decode packet")]
-    DecodeFail(#[from] DecodeError),
 }
 
 impl Socket {
@@ -361,216 +300,4 @@ fn spawn_inbound_peer_task(
             }
         }
     });
-}
-
-/// A UDP packet sent over the network. These packets have the following format:
-///
-/// A header, consisting of:
-/// - 4 bytes: Magic number (0x010203)
-/// - 1 byte: Packet type (0 = SYN, 1 = ACK, 2 = SYNACK, 3 = HEARTBEAT, 4 = DATA) - see [SocketPacketType].
-/// - 4 bytes: Sequence number
-/// - 4 bytes: Length of the compressed data
-/// - 4 bytes: Length of the uncompressed data
-///
-/// Then arbitrary-length data, as defined by the protocol.
-#[derive(Debug, Clone)]
-pub struct SocketPacket {
-    /// The type of packet.
-    pub packet_type: SocketPacketType,
-    /// The sequence of the underlying [ProtocolPacket].
-    pub packet_number: u32,
-    /// The chunk number of the packet. This is only used for data packets.
-    pub chunk_number: u32,
-    /// The length of the packet, compressed.
-    pub compressed_data_length: u32,
-    /// The length of the packet
-    pub uncompressed_data_length: u32,
-    /// The packet data. This is empty for SYN, ACK, SYNACK, and HEARTBEAT packets.
-    pub compressed_data: Vec<u8>,
-}
-
-impl PartialEq for SocketPacket {
-    fn eq(&self, other: &Self) -> bool {
-        self.packet_type == other.packet_type
-            && self.packet_number == other.packet_number
-            && self.chunk_number == other.chunk_number
-    }
-}
-
-impl Eq for SocketPacket {}
-
-// TODO: Justify that packet_number and chunk_number will be unique.
-
-impl PartialOrd for SocketPacket {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        //  enforces lexicographic ordering, with packet number taking precedence
-        if self.packet_number == other.packet_number {
-            Some(self.chunk_number.cmp(&other.chunk_number))
-        } else {
-            Some(self.packet_number.cmp(&other.packet_number))
-        }
-    }
-}
-
-impl Ord for SocketPacket {
-    fn cmp(&self, other: &Self) -> Ordering {
-        if self.packet_number == other.packet_number {
-            self.chunk_number.cmp(&other.chunk_number)
-        } else {
-            self.packet_number.cmp(&other.packet_number)
-        }
-    }
-}
-
-/// An enumeration of the different types of network packets.
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-#[repr(u8)]
-pub enum SocketPacketType {
-    /// Packets sent by the initiating peer.
-    Syn,
-    /// Packets sent by a receiving peer.
-    Ack,
-    /// Packets sent by the initiating peer after receiving an ACK. Once this is sent, the connection is established.
-    SynAck,
-    /// Packets sent by either peer to keep the connection alive. This is done to avoid stateful firewalls from dropping the connection.
-    Heartbeat,
-    /// Actual communication data
-    Data,
-    /// An invalid packet.
-    Invalid,
-}
-
-impl From<u8> for SocketPacketType {
-    fn from(value: u8) -> Self {
-        match value {
-            0 => SocketPacketType::Syn,
-            1 => SocketPacketType::Ack,
-            2 => SocketPacketType::SynAck,
-            3 => SocketPacketType::Heartbeat,
-            4 => SocketPacketType::Data,
-            _ => SocketPacketType::Invalid,
-        }
-    }
-}
-
-impl SocketPacket {
-    /// Create a new packet with the given type, sequence number, and data.
-    pub fn new<Data>(
-        packet_type: SocketPacketType,
-        packet_number: u32,
-        chunk_number: u32,
-        data: Data,
-    ) -> io::Result<Self>
-    where
-        Data: AsRef<[u8]>,
-    {
-        let data = data.as_ref();
-        let mut encoder = GzEncoder::new(vec![], Compression::default());
-        let _ = encoder.write_all(data);
-        let compressed_data = encoder.finish()?;
-
-        Ok(Self {
-            packet_type,
-            packet_number,
-            chunk_number,
-            uncompressed_data_length: data.len() as u32,
-            compressed_data_length: compressed_data.len() as u32,
-            compressed_data,
-        })
-    }
-
-    /// Encode the packet into a byte buffer.
-    pub fn encode(&self) -> io::Result<Vec<u8>> {
-        let mut buf = Vec::with_capacity(MIN_SOCKET_PACKET_SIZE);
-
-        // write header
-        buf.write_u24::<BigEndian>(SOCKET_PACKET_MAGIC_NUMBER)?;
-        buf.write_u8(self.packet_type as u8)?;
-        buf.write_u32::<BigEndian>(self.packet_number)?;
-        buf.write_u32::<BigEndian>(self.chunk_number)?;
-        buf.write_u32::<BigEndian>(self.compressed_data_length)?;
-        buf.write_u32::<BigEndian>(self.uncompressed_data_length)?;
-
-        // write data
-        buf.write_all(&self.compressed_data)?;
-
-        Ok(buf)
-    }
-
-    /// Create an empty packet with the given type, sequence number, and chunk number. Useful for non-data packets.
-    pub fn empty(packet_type: SocketPacketType, packet_number: u32, chunk_number: u32) -> Self {
-        Self {
-            packet_type,
-            packet_number,
-            chunk_number,
-            compressed_data: vec![],
-            compressed_data_length: 0,
-            uncompressed_data_length: 0,
-        }
-    }
-
-    /// Decode a packet from the given byte buffer.
-    pub fn decode<Data>(bytes: Data) -> Result<SocketPacket, SocketPacketDecodeError>
-    where
-        Data: AsRef<[u8]>,
-    {
-        let bytes = bytes.as_ref();
-
-        // check minimum packet length
-        if bytes.len() < MIN_SOCKET_PACKET_SIZE {
-            return Err(SocketPacketDecodeError::BadSize);
-        }
-
-        // create reader
-        let mut reader = Cursor::new(bytes);
-
-        // check magic number
-        let magic = reader.read_u24::<BigEndian>()?;
-        if magic != SOCKET_PACKET_MAGIC_NUMBER {
-            return Err(SocketPacketDecodeError::BadMagic);
-        }
-
-        // read packet header
-        let packet_type = reader.read_u8()?.into();
-        let packet_number = reader.read_u32::<BigEndian>()?;
-        let chunk_number = reader.read_u32::<BigEndian>()?;
-        let compressed_data_length = reader.read_u32::<BigEndian>()?;
-        let uncompressed_data_length = reader.read_u32::<BigEndian>()?;
-
-        if (uncompressed_data_length as usize) == 0 {
-            return Ok(SocketPacket::new(
-                packet_type,
-                packet_number,
-                chunk_number,
-                vec![],
-            )?);
-        }
-
-        // read data
-        let mut data = vec![0; compressed_data_length as usize];
-        reader.read_exact(&mut data)?;
-
-        Ok(SocketPacket::new(
-            packet_type,
-            packet_number,
-            chunk_number,
-            data,
-        )?)
-    }
-
-    /// Decompress the data.
-    pub fn decompress(&self) -> io::Result<Vec<u8>> {
-        let mut data = vec![0; self.uncompressed_data_length as usize];
-        let mut gz_decoder = GzDecoder::new(self.compressed_data.as_slice());
-        gz_decoder.read_exact(&mut data)?;
-        Ok(data)
-    }
-}
-
-impl TryFrom<SocketPacket> for ProtocolPacket {
-    type Error = SocketPacketDecodeError;
-
-    fn try_from(value: SocketPacket) -> Result<Self, Self::Error> {
-        Ok(try_decode_packet(value.decompress()?)?)
-    }
 }
